Data augmentation, learning transfer and dropout
In this script I use data augmentation using ImageDataGenerator from Keras to generate a training set and a validation set to train a neural network.
This neural network consists in two parts. First, several pretrained layers from InceptionV3 (in Keras) which cannot be trained (since in the script I set layer.trainable as False. Second, one Dense layer with 1024 neurons and activation ReLU and one output layer with one neuron and activation sigmoid to perform binary classification between horses and humans. At the end of this script I play with the Dropout, cutting the connection of 20% of neurons from the Dense layer with 1024 units to the output layer.